pyproject.toml
[tool.poetry]
name = "ppi-repository"
version = "0.1.0"
description = "Deep Learning Models for Protein-Protein-Interaction (PPI) Prediction."
authors = ["Johannes Melsbach <github@melsbach.org>"]
readme = "README.md"
packages = [{include = "ppi", from="src"}]

[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.31.0"
pandas = "^2.0.3"
numpy = "^1.25.2"
lightning = "^2.0.7"
torchmetrics = "^1.0.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
